Add install-images dependency on install-target
In current buildroot, install-target has an intra-package dependency
on install-staging to ensure that staging artifacts can be consumed
when installing to target.

Because of the way the boot/syslinux package is constructed, it
requires a similar dependency to ensure that target artifacts can be
consumed when installing images.

Without this dependency, a race condition is apparent when running a
parallel build, which can result in intermittent failures.

This patch should be upstreamed, if I can figure out how to do that.